# Conflicts:
#	.github/workflows/ci.yml
#	.github/workflows/sentry-release.yml
#	.github/workflows/update-web-mod-definitions.yml
#	osu.Game.Tournament.Tests/Screens/TestSceneGameplayScreen.cs
#	osu.Game.Tournament.Tests/Screens/TestSceneSeedingScreen.cs
#	osu.Game.Tournament/Components/SongBar.cs
#	osu.Game.Tournament/Components/TournamentMatchChatDisplay.cs
#	osu.Game.Tournament/Models/TournamentBeatmap.cs
#	osu.Game.Tournament/Screens/Gameplay/GameplayScreen.cs
#	osu.Game/OsuGame.cs
#	osu.Game/Overlays/NotificationOverlay.cs
#	osu.Game/Screens/OnlinePlay/Multiplayer/Spectate/MultiSpectatorScreen.cs
#	osu.Game/Screens/OnlinePlay/Multiplayer/Spectate/PlayerGrid.Cell.cs
#	osu.Game/Screens/OnlinePlay/Multiplayer/Spectate/PlayerGrid.cs
#	osu.Game/Screens/Play/HUDOverlay.cs
#	osu.Game/Screens/Play/SpectatorPlayer.cs
#	osu.Game/Screens/Ranking/Expanded/Accuracy/RankText.cs
#	osu.Game/Screens/Select/Carousel/DrawableCarouselBeatmap.cs

+++ osu.Game.Tournament/Screens/Gameplay/Components/TeamScore.cs
@@ -17,10 +17,10 @@ namespace osu.Game.Tournament.Screens.Gameplay.Components
 {
     public partial class TeamScore : CompositeDrawable
     {
-        private readonly Bindable<int?> currentTeamScore = new Bindable<int?>();
+        private readonly Bindable<long?> currentTeamScore = new Bindable<long?>();
         private readonly StarCounter counter;
 
-        public TeamScore(Bindable<int?> score, TeamColour colour, int count)
+        public TeamScore(Bindable<long?> score, TeamColour colour, int count)
         {
             bool flip = colour == TeamColour.Blue;
             var anchor = flip ? Anchor.TopRight : Anchor.TopLeft;
@@ -37,7 +37,7 @@ namespace osu.Game.Tournament.Screens.Gameplay.Components
             currentTeamScore.BindTo(score);
         }
 
-        private void scoreChanged(ValueChangedEvent<int?> score) => counter.Current = score.NewValue ?? 0;
+        private void scoreChanged(ValueChangedEvent<long?> score) => counter.Current = score.NewValue ?? 0;
 
         public partial class TeamScoreStarCounter : StarCounter
         {
